+++ src/main/java/com/example/apipeticos/repositories/DayHintRepository.java
@@ -3,7 +3,7 @@ package com.example.apipeticos.repositories;
 import com.example.apipeticos.models.DayHint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DayHintRepository extends JpaRepository<DayHint, Long> {
+public interface DayHintRepository extends JpaRepository<DayHint, Integer> {
 
 
 }
